requirements/test.txt
#
# This file is autogenerated by pip-compile with python 3.8
# To update, run:
#
#    make upgrade
#
anyio==3.5.0
    # via starlette
appdirs==1.4.4
    # via fs
asgiref==3.5.0
    # via
    #   django
    #   uvicorn
attrs==21.4.0
    # via pytest
boto==2.49.0
    # via -r requirements/base.in
certifi==2021.10.8
    # via requests
cffi==1.15.0
    # via cryptography
chardet==4.0.0
    # via pysrt
charset-normalizer==2.0.11
    # via requests
click==8.0.3
    # via
    #   code-annotations
    #   pact-python
    #   uvicorn
code-annotations==1.3.0
    # via edx-toggles
coverage[toml]==6.3.1
    # via
    #   -r requirements/test.in
    #   pytest-cov
cryptography==36.0.1
    # via
    #   djfernet
    #   pyjwt
ddt==1.4.4
    # via -r requirements/test.in
    # via
    #   -c requirements/common_constraints.txt
    #   -r requirements/base.in
    #   django-crum
    #   django-model-utils
    #   django-storages
    #   djangorestframework
    #   drf-jwt
    #   edx-django-utils
    #   edx-drf-extensions
    #   edx-toggles
django-crum==0.7.9
    # via
    #   edx-django-utils
    #   edx-toggles
django-model-utils==4.2.0
    # via -r requirements/base.in
django-storages==1.12.3
    # via -r requirements/base.in
django-waffle==2.3.0
    # via
    #   edx-django-utils
    #   edx-drf-extensions
    #   edx-toggles
djangorestframework==3.13.1
    # via
    #   drf-jwt
    #   edx-drf-extensions
djfernet==0.7.4
    # via -r requirements/base.in
drf-jwt==1.19.2
    # via edx-drf-extensions
edx-django-utils==4.5.0
    # via
    #   edx-drf-extensions
    #   edx-toggles
edx-drf-extensions==8.0.1
    # via -r requirements/base.in
edx-opaque-keys==2.3.0
    # via edx-drf-extensions
edx-toggles==4.3.1
    # via -r requirements/base.in
fastapi==0.73.0
    # via pact-python
fs==2.4.15
    # via -r requirements/test.in
future==0.18.2
    # via pyjwkest
h11==0.13.0
    # via uvicorn
idna==3.3
    # via
    #   anyio
    #   requests
iniconfig==1.1.1
    # via pytest
jinja2==3.0.3
    # via code-annotations
lxml==4.7.1
    # via -r requirements/base.in
markupsafe==2.0.1
    # via jinja2
mock==4.0.3
    # via -r requirements/test.in
newrelic==7.4.0.172
    # via edx-django-utils
packaging==21.3
    # via pytest
pact-python==1.5.0
    # via -r requirements/test.in
pbr==5.8.1
    # via stevedore
pluggy==1.0.0
    # via pytest
psutil==5.9.0
    # via
    #   edx-django-utils
    #   pact-python
py==1.11.0
    # via pytest
pycparser==2.21
    # via cffi
pycryptodomex==3.14.1
    # via pyjwkest
pydantic==1.9.0
    # via fastapi
pyjwkest==1.4.2
    # via edx-drf-extensions
pyjwt[crypto]==2.3.0
    # via
    #   drf-jwt
    #   edx-drf-extensions
pymongo==3.12.3
    # via edx-opaque-keys
pyparsing==3.0.7
    # via packaging
pysrt==1.1.2
    # via -r requirements/base.in
pytest==7.0.0
    # via
    #   pytest-cov
    #   pytest-django
pytest-cov==3.0.0
    # via -r requirements/test.in
pytest-django==4.5.2
    # via -r requirements/test.in
python-dateutil==2.8.2
    # via edx-drf-extensions
python-slugify==5.0.2
    # via code-annotations
pytz==2021.3
    # via
    #   django
    #   djangorestframework
    #   fs
pyyaml==6.0
    # via code-annotations
requests==2.27.1
    # via
    #   edx-drf-extensions
    #   pact-python
    #   pyjwkest
    #   responses
responses==0.18.0
    # via -r requirements/test.in
semantic-version==2.9.0
    # via edx-drf-extensions
six==1.16.0
    # via
    #   edx-drf-extensions
    #   fs
    #   pact-python
    #   pyjwkest
    #   python-dateutil
sniffio==1.2.0
    # via anyio
sqlparse==0.4.2
    # via django
starlette==0.17.1
    # via fastapi
stevedore==3.5.0
    # via
    #   code-annotations
    #   edx-django-utils
    #   edx-opaque-keys
text-unidecode==1.3
    # via python-slugify
tomli==2.0.0
    # via
    #   coverage
    #   pytest
typing-extensions==4.0.1
    # via pydantic
urllib3==1.26.8
    # via
    #   pact-python
    #   requests
    #   responses
uvicorn==0.17.4
    # via pact-python

# The following packages are considered to be unsafe in a requirements file:
# setuptools
